jghj

+++ pp2.5/src/pp2/pkg5/Pp25.java
@@ -3,10 +3,10 @@ import java.util.Scanner;
 
 public class Pp25 {
     public static void main(String[] args) {
-        float temperature, finalTemp;
+        double temperature, finalTemp;
         Scanner in = new Scanner(System.in);
         System.out.print("Enter temperature in Fahrenheit: ");
-        temperature = in.nextInt();
+        temperature = in.nextDouble();
         finalTemp = ((temperature - 32)*5)/9;
         System.out.println("temperature in Celsius = " + finalTemp);
     }
